requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
alembic==1.12.0
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0
jinja2==3.1.2
cryptography==41.0.4
python-dateutil==2.8.2
pydantic-settings==2.0.1